Add : 모임방 공지(방장 권한 추가)

+++ src/main/java/com/github/gather/service/NoticeService.java
@@ -4,6 +4,7 @@ import com.github.gather.dto.NoticeDto;
 import com.github.gather.entity.GroupMember;
 import com.github.gather.entity.GroupTable;
 import com.github.gather.entity.NoticeEntity;
+import com.github.gather.entity.Role.GroupMemberRole;
 import com.github.gather.entity.User;
 import com.github.gather.exception.ExceptionMessage;
 import com.github.gather.repositroy.NoticeRepository;
@@ -18,20 +19,26 @@ import java.util.stream.Collectors;
 @Service
 public class NoticeService {
 
-
     private final NoticeRepository noticeRepository;
     private final AuthService authService;
     private final GroupMemberRepository groupMemberRepository;
+    private final GroupTableService groupTableService;
 
-    public NoticeService(NoticeRepository noticeRepository, AuthService authService, GroupMemberRepository groupMemberRepository) {
+    public NoticeService(NoticeRepository noticeRepository, AuthService authService, GroupMemberRepository groupMemberRepository, GroupTableService groupTableService) {  // 수정
         this.noticeRepository = noticeRepository;
         this.authService = authService;
         this.groupMemberRepository = groupMemberRepository;
+        this.groupTableService = groupTableService;
     }
 
     // 공지 등록
     public NoticeDto createNotice(GroupTable groupId, NoticeDto noticeDto, HttpServletRequest request) {
         User user = authService.checkToken(request);
+        GroupMember groupMember = groupMemberRepository.findByUserIdAndGroupId(user, groupId)
+                .orElseThrow(() -> new IllegalArgumentException(ExceptionMessage.USER_NOT_PART_OF_GROUP));
+        if (groupMember.getRole() != GroupMemberRole.LEADER) {
+            throw new IllegalArgumentException(ExceptionMessage.ONLY_LEADER_CAN_POST_NOTICE);
+        }
 
         NoticeEntity notice = NoticeEntity.builder()
                 .title(noticeDto.getTitle())
@@ -51,8 +58,9 @@ public class NoticeService {
     }
 
     // 공지 전체 조회
-    public List<NoticeDto> getAllNotices() {
-        List<NoticeEntity> noticeEntities = noticeRepository.findAll();
+    public List<NoticeDto> getAllNoticesByGroup(Long groupId) {
+        GroupTable group = groupTableService.getGroupById(groupId);
+        List<NoticeEntity> noticeEntities = noticeRepository.findByGroupId(group);
         return noticeEntities.stream()
                 .map(notice -> NoticeDto.builder()
                         .noticeIdx(notice.getNoticeIdx())
@@ -80,7 +88,17 @@ public class NoticeService {
     // 공지 수정
     public NoticeDto updateNotice(Long noticeIdx, NoticeDto noticeDto, HttpServletRequest request) {
         User user = authService.checkToken(request);
-        NoticeEntity notice = noticeRepository.findById(noticeIdx).get();
+
+        NoticeEntity notice = noticeRepository.findById(noticeIdx).orElseThrow(() -> new IllegalArgumentException(ExceptionMessage.NOTICE_NOT_FOUND));
+        GroupTable groupTable = notice.getGroupId();
+
+        GroupMember groupMember = groupMemberRepository.findByUserIdAndGroupId(user, groupTable)
+                .orElseThrow(() -> new IllegalArgumentException(ExceptionMessage.USER_NOT_PART_OF_GROUP));
+
+        if (groupMember.getRole() != GroupMemberRole.LEADER) {
+            throw new IllegalArgumentException(ExceptionMessage.NO_AUTH_TO_MODIFY_NOTICE);
+        }
+
         NoticeEntity updatedNotice = noticeRepository.save(NoticeEntity.builder()
                 .noticeIdx(noticeIdx)
                 .title(noticeDto.getTitle())
@@ -89,6 +107,7 @@ public class NoticeService {
                 .groupId(notice.getGroupId())
                 .userId(user)
                 .build());
+
         return NoticeDto.builder()
                 .noticeIdx(updatedNotice.getNoticeIdx())
                 .title(updatedNotice.getTitle())
@@ -98,10 +117,21 @@ public class NoticeService {
                 .build();
     }
 
+
     // 공지 삭제
     public void deleteNotice(Long noticeIdx, HttpServletRequest request) {
         User user = authService.checkToken(request);
-        NoticeEntity notice = noticeRepository.findById(noticeIdx).get();
+
+        NoticeEntity notice = noticeRepository.findById(noticeIdx).orElseThrow(() -> new IllegalArgumentException(ExceptionMessage.NOTICE_NOT_FOUND));
+        GroupTable groupTable = notice.getGroupId();
+
+        GroupMember groupMember = groupMemberRepository.findByUserIdAndGroupId(user, groupTable)
+                .orElseThrow(() -> new IllegalArgumentException(ExceptionMessage.USER_NOT_PART_OF_GROUP));
+
+        if (groupMember.getRole() != GroupMemberRole.LEADER) {
+            throw new IllegalArgumentException(ExceptionMessage.NO_AUTH_TO_DELETE_NOTICE);
+        }
+
         noticeRepository.deleteById(noticeIdx);
     }
 }
